# Fix editoast batch simulation reducing DB queries