--- a/editoast/src/views/v2/timetable.rs
+++ b/editoast/src/views/v2/timetable.rs
@@ -2,6 +2,7 @@ pub mod stdcm;
 
 use std::collections::HashMap;
 use std::ops::DerefMut as _;
+use std::sync::Arc;
 
 use actix_web::delete;
 use actix_web::get;
@@ -34,11 +35,13 @@ use crate::modelsv2::timetable::TimetableWithTrains;
 use crate::modelsv2::train_schedule::TrainSchedule;
 use crate::modelsv2::train_schedule::TrainScheduleChangeset;
 use crate::modelsv2::Infra;
+use crate::modelsv2::RollingStockModel;
 use crate::views::pagination::PaginatedList;
 use crate::views::pagination::PaginationQueryParam;
 use crate::views::pagination::PaginationStats;
 use crate::views::v2::train_schedule::train_simulation_batch;
 use crate::views::v2::train_schedule::TrainScheduleForm;
+use crate::views::v2::train_schedule::TrainScheduleProxy;
 use crate::views::v2::train_schedule::TrainScheduleResult;
 use crate::CoreClient;
 use crate::RedisClient;
@@ -333,17 +336,29 @@ pub async fn conflicts(
     let infra_id = query.into_inner().infra_id;
 
     // 1. Retrieve Timetable / Infra / Trains / Simultion
-    let timetable = TimetableWithTrains::retrieve_or_fail(conn, timetable_id, || {
+    let timetable_trains = TimetableWithTrains::retrieve_or_fail(conn, timetable_id, || {
         TimetableError::NotFound { timetable_id }
     })
     .await?;
+    let timetable: Timetable = timetable_trains.clone().into();
 
     let infra = Infra::retrieve_or_fail(conn, infra_id, || TimetableError::InfraNotFound {
         infra_id,
     })
     .await?;
 
-    let (trains, _): (Vec<_>, _) = TrainSchedule::retrieve_batch(conn, timetable.train_ids).await?;
+    let (trains, _): (Vec<_>, _) =
+        TrainSchedule::retrieve_batch(conn, timetable_trains.train_ids).await?;
+
+    let (rolling_stocks, _): (Vec<_>, _) = RollingStockModel::retrieve_batch(
+        db_pool.get().await?.deref_mut(),
+        trains
+            .iter()
+            .map::<String, _>(|t| t.rolling_stock_name.clone()),
+    )
+    .await?;
+
+    let proxy = Arc::new(TrainScheduleProxy::new(&rolling_stocks, &[timetable]));
 
     let simulations = train_simulation_batch(
         db_pool.clone(),
@@ -351,6 +366,7 @@ pub async fn conflicts(
         core_client.clone(),
         &trains,
         &infra,
+        proxy,
     )
     .await?;