We reduce the number of queries executed by the batch simulation function. 
Previously, for `N` trains, we executed `Nx4` postgres queries. 
Now we run ~4 batch queries.

close #7867 

close #7907

> [!NOTE]
> In my test in local the health check takes no more than 500ms to respond, compared with 6s before 

## Benchmark

I tested with a timetable of 823 trains

### Before

**Summary**:

- No cache: 20.96s
- Redis cache: 1.66s

**Project path**:

- No cache: 27.25s
- Redis cache: 3.69s

### After

**Summary**:

- No cache: 22s
- Redis cache: 964ms

**Project path**:

- No cache: 26s
- Redis cache: 3s